+++ lua/hygen/compat.lua
@@ -1,0 +1,9 @@
+local M = {}
+
+-- see changed features in treesitter section
+-- https://neovim.io/doc/user/news-0.10.html#_changed-features
+M.has_v_0_10 = vim.fn.has("nvim-0.10") == 1
+
+M.predicate_options = M.has_v_0_10 and {} or nil
+
+return M
