+++ lua/hygen/tree-sitter/init.lua
@@ -1,0 +1,9 @@
+local M = {}
+
+function M.setup()
+  require("hygen.tree-sitter.parser-info").setup()
+  require("hygen.tree-sitter.predicates").setup()
+  require("hygen.tree-sitter.directives").setup()
+end
+
+return M
